Fix CI failures on ANSSI compliance PR.
Resolve clippy pedantic/nursery violations, move repro test modules after
main code, bump rand to 0.10 and refresh lockfile for audit (time/RUSTSEC),
and make the outdated workflow informational on pull requests.

+++ src/send/udp.rs
@@ -21,7 +21,7 @@ pub fn start<C>(sender: &send::Sender<C>) -> Result<(), send::Error> {
     let sock_buffer_size = sock_utils::get_socket_send_buffer_size(&socket)?;
     log::info!("UDP socket send buffer size set to {sock_buffer_size}");
 
-    if (sock_buffer_size as i32) < buffer_size {
+    if sock_buffer_size < buffer_size {
         log::warn!(
             "UDP socket send buffer may be too small ({sock_buffer_size} < {buffer_size}) to achieve optimal performances"
         );
